+++ packages/next/test/integration/ppr-root-params/next.config.js
@@ -1,6 +1,3 @@
 module.exports = {
-  experimental: {
-    cacheComponents: true,
-    dynamicIO: true,
-  },
+  cacheComponents: true,
 };